# public/swagger.yaml
swagger: "2.0"
info:
  title: 게시판
paths:
...
  /api/board/createPost:
    post:
      summary: (글 생성)
      tags:
        - board
      parameters:
        - name: body
          in: body
          description: "생성한 글 정보"
          required: true
          schema:
            $ref: '#/board/createPost/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/board/createPost/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/board/response_400'
  /api/board/requestBoardForm:
    post:
      summary: (게시글 요청)
      tags:
        - board
      parameters:
        - name: body
          in: body
          description: "게시글 정보"
          required: true
          schema:
            $ref: '#/board/requestBoardForm/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/board/requestBoardForm/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/board/response_400'
  /api/board/deletePost:
    post:
      summary: (게시글 삭제 요청)
      tags:
        - board
      parameters:
        - name: body
          in: body
          description: "게시글 정보"
          required: true
          schema:
            $ref: '#/board/deletePost/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/board/deletePost/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/board/response_400'
  /api/board/modifyPost:
    post:
      summary: (게시글 내용 수정)
      tags:
        - board
      parameters:
        - name: body
          in: body
          description: "게시글 수정 사항"
          required: true
          schema:
            $ref: '#/board/modifyPost/request'

      responses:
        '이전 게시글 내용과 동일함 200':
          description: "성공"
          schema:
            $ref: '#/board/modifyPost/response_200_case_1'

        '게시글 수정 완료 200':
          description: "성공"
          schema:
            $ref: '#/board/modifyPost/response_200_case_2'

        '400':
          description: "에러"
          schema:
            $ref: '#/board/response_400'
  /api/board/getTotal:
    post:
      summary: (전체 게시물의 갯수)
      tags:
        - board
 
      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/board/getTotal/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/board/response_400'
  /api/board/getFavorite:
    post:
      summary: (유저가 좋아요 정보를 눌렀는지 확인 요청)
      tags:
        - board
      parameters:
        - name: body
          in: body
          description: "해당 게시글 정보"
          required: true
          schema:
            $ref: '#/board/getFavorite/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/board/getFavorite/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/board/response_400'
  /api/board/favorite:
    post:
      summary: (좋아요 누름 요청)
      tags:
        - board
      parameters:
        - name: body
          in: body
          description: "좋아요 누른 게시글 정보"
          required: true
          schema:
            $ref: '#/board/favorite/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/board/favorite/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/board/response_400'
  /api/board/unFavorite:
    post:
      summary: (좋아요 취소 요청)
      tags:
        - board
      parameters:
        - name: body
          in: body
          description: "좋아요 취소한 게시글 정보"
          required: true
          schema:
            $ref: '#/board/unFavorite/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/board/unFavorite/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/board/response_400'


##### board 모델 #####
...